+++ tests-integration/tests/cpc_test/main.rs
@@ -16,7 +16,6 @@
 // under the License.
 
 mod deserialize;
-mod summary;
 mod union;
 mod update;
 mod wrapper;