+++ tests-integration/tests/hll_test/main.rs
@@ -16,6 +16,5 @@
 // under the License.
 
 mod bounds;
-mod summary;
 mod union;
 mod update;